ap/services/ingestion/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pandas
openpyxl
python-multipart
pymupdf
python-docx
odfpy
